requirements.txt
homeabsl-py
accelerate 
aiohttp 
aiosignal 
attrs 
audioread 
Bottleneck 
Brotli 
certifi 
cffi 
charset-normalizer 
contourpy 
cycler 
datasets 
decorator 
dill 
einops 
filelock 
flash_attn 
flatbuffers
fonttools 
frozenlist 
fsspec 
gmpy2 
huggingface_hub 
idna 
imageio 
imageio-ffmpeg 
importlib-metadata 
jax
jaxlib
Jinja2 
joblib 
kiwisolver 
lazy_loader 
librosa 
llvmlite 
MarkupSafe 
matplotlib
mediapipe
mkl-fft 
mkl-random 
mkl-service
ml-dtypes
moviepy 
mpmath 
msgpack 
multidict 
multiprocess 
munkres
networkx 
numba 
numexpr 
numpy 
opencv-contrib-python
opt-einsum
packaging 
pandas 
pillow 
platformdirs 
pooch 
proglog
protobuf
psutil 
pyarrow 
pyarrow-hotfix 
pycparser 
pyparsing 
PySocks 
python-dateutil 
pytz 
PyYAML 
regex 
requests 
safetensors 
scikit-learn 
scipy 
six 
sounddevice
soundfile 
soxr 
sympy 
threadpoolctl 
tokenizers 
torch 
torchvision 
tqdm 
transformers 
triton 
typing_extensions 
tzdata 
urllib3 
xxhash 
yarl 
zipp 